+++ system/application_manager/app_manager.c
@@ -14,7 +14,7 @@
 #include "../watchdog_timer/watchdog_timer.h"
 #include <src_user/TlmCmd/command_definitions.h>
 #include "../../library/print.h"
-#include "../../TlmCmd/common_cmd_packet_util.h"
+#include "../../tlm_cmd/common_cmd_packet_util.h"
 
 static AM_ACK AM_initialize_app_(size_t id);
 static AM_ACK AM_execute_app_(size_t id);
