+++ system/application_manager/app_manager.h
@@ -8,7 +8,7 @@
 
 #include <stddef.h>
 #include "app_info.h"
-#include "../../TlmCmd/common_cmd_packet.h"
+#include "../../tlm_cmd/common_cmd_packet.h"
 
 #define AM_TLM_PAGE_SIZE (32)                               //!< AMのAppInfoテーブルの1テレメトリパケット(=1ページ)に格納されるAppInfo数（ページネーション用）
 #define AM_TLM_PAGE_MAX (4)                                 //!< AMのAppInfoテーブルのページ数（ページネーション用）
