+++ system/event_manager/event_handler.c
@@ -7,7 +7,7 @@
 #include "event_handler.h"
 #include <string.h>
 #include <stdlib.h>
-#include "../../TlmCmd/common_cmd_packet_util.h"
+#include "../../tlm_cmd/common_cmd_packet_util.h"
 #include "../../applications/timeline_command_dispatcher_id_define.h"
 #include "../time_manager/time_manager.h"
