+++ system/mode_manager/mode_manager.c
@@ -10,10 +10,10 @@
 #include "../time_manager/time_manager.h"
 #include "../task_manager/task_dispatcher.h"
 #include "../event_manager/event_logger.h"
-#include "../../TlmCmd/block_command_executor.h"
-#include "../../TlmCmd/common_cmd_packet_util.h"
+#include "../../tlm_cmd/block_command_executor.h"
+#include "../../tlm_cmd/common_cmd_packet_util.h"
 #include <src_user/TlmCmd/command_definitions.h>
-#include "../../TlmCmd/packet_handler.h"
+#include "../../tlm_cmd/packet_handler.h"
 #include "../../applications/timeline_command_dispatcher_id_define.h"
 #include "../../library/endian.h"
