+++ system/task_manager/task_dispatcher.c
@@ -10,9 +10,9 @@
 #include <stdlib.h>
 #include <string.h> // for memcpy
 
-#include "../../TlmCmd/packet_list_util.h"
-#include "../../TlmCmd/block_command_executor.h"
-#include "../../TlmCmd/common_cmd_packet_util.h"
+#include "../../tlm_cmd/packet_list_util.h"
+#include "../../tlm_cmd/block_command_executor.h"
+#include "../../tlm_cmd/common_cmd_packet_util.h"
 #include "../mode_manager/mode_manager.h"
 #include "../time_manager/time_manager.h"
 #include "../time_manager/obc_time_config.h"
