+++ system/time_manager/time_manager.h
@@ -6,7 +6,7 @@
 #define TIME_MANAGER_H_
 
 #include "obc_time.h"
-#include "../../TlmCmd/common_cmd_packet.h"
+#include "../../tlm_cmd/common_cmd_packet.h"
 
 #define TMGR_DEFAULT_UNIXTIME_EPOCH_FOR_UTL (1577836800.0)  /*!< 2020-01-01T00:00:00Z 時点の unixtime.
                                                                  utl_unixtime_epoch のデフォルト値 */
