+++ system/watchdog_timer/watchdog_timer.c
@@ -6,7 +6,7 @@
  */
 #include "watchdog_timer.h"
 #include "../../library/print.h"
-#include "../../TlmCmd/common_cmd_packet_util.h"
+#include "../../tlm_cmd/common_cmd_packet_util.h"
 
 static WDT_Config wdt_config_;
 const WDT_Config* const wdt_config = &wdt_config_;
