+++ system/watchdog_timer/watchdog_timer.h
@@ -8,7 +8,7 @@
 #define WATCHDOG_TIMER_H_
 
 #include "../../hal/wdt.h"
-#include "../../TlmCmd/common_cmd_packet.h"
+#include "../../tlm_cmd/common_cmd_packet.h"
 
 extern const WDT_Config* const wdt_config;
